+++ source/app/service-providers/windows/create-project-properties-window.ts
@@ -45,10 +45,7 @@ export default function createProjectPropertiesWindow (logger: LogProvider, conf
     show: false,
     fullscreenable: false,
     webPreferences: {
-      // contextIsolation and sandbox mean: Preload scripts have access to
-      // Node modules, the renderers not
-      contextIsolation: true,
-      sandbox: false,
+      sandbox: true,
       preload: PROJECT_PROPERTIES_PRELOAD_WEBPACK_ENTRY
     }
   }