+++ source/app/service-providers/windows/create-tag-manager-window.ts
@@ -43,10 +43,7 @@ export default function createTagManagerWindow (logger: LogProvider, config: Con
     show: false,
     fullscreenable: false,
     webPreferences: {
-      // contextIsolation and sandbox mean: Preload scripts have access to
-      // Node modules, the renderers not
-      contextIsolation: true,
-      sandbox: false,
+      sandbox: true,
       preload: TAG_MANAGER_PRELOAD_WEBPACK_ENTRY
     }
   }